pyproject.toml
[tool.poetry]
name = "takehome"
version = "0.1.0"
description = "Supersweet ML/Data Engineer Take Home"
authors = ["Jeremy Greer <jeremy.greer@anthem.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.2"
pyarrow = "^5.0.0"
"fhir.resources" = "^6.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.12.1"
mypy = "^0.910"
isort = "^5.9.3"
black = {extras = ["jupyter"], version = "^21.8b0"}
flake8 = "^3.9.2"
pydocstyle = "^6.1.1"
ipykernel = "^6.3.1"
pytest-mock = "^3.6.1"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--tb=short"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "nocov",
    "if TYPE_CHECKING:",
    ]
    
[tool.coverage.run]
# Activating branch coverage is super important
branch = true
omit = [
  # add your files to omit here
    ]
